Started implementation of Meta Data Table parsing

+++ PeNet.UnitTest/Utilities/HeapOffsetBasedIndexSizes_Test.cs
@@ -28,7 +28,7 @@ namespace PeNet.UnitTest.Utilities
         [InlineData(0x01, 4)]
         [InlineData(0x21, 4)]
         [InlineData(0x31, 4)]
-        public void StringIndexSize_Test(byte heapOffsetSizes, int result)
+        public void StringIndexSize_Test(byte heapOffsetSizes, uint result)
         {
             var heapOffsetBasedIndexSizes = new PeNet.Utilities.HeapOffsetBasedIndexSizes(heapOffsetSizes);
             Assert.Equal(result, heapOffsetBasedIndexSizes.StringIndexSize);
@@ -41,7 +41,7 @@ namespace PeNet.UnitTest.Utilities
         [InlineData(0x02, 4)]
         [InlineData(0x03, 4)]
         [InlineData(0x42, 4)]
-        public void GuidIndexSizeShouldBeTwo_Test(byte heapOffsetSizes, int result)
+        public void GuidIndexSizeShouldBeTwo_Test(byte heapOffsetSizes, uint result)
         {
             var heapOffsetBasedIndexSizes = new PeNet.Utilities.HeapOffsetBasedIndexSizes(heapOffsetSizes);
             Assert.Equal(result, heapOffsetBasedIndexSizes.GuidIndexSize);
@@ -54,7 +54,7 @@ namespace PeNet.UnitTest.Utilities
         [InlineData(0x04, 4)]
         [InlineData(0x54, 4)]
         [InlineData(0x05, 4)]
-        public void BlobIndexSizeShouldBeTwo_Test(byte heapOffsetSizes, int result)
+        public void BlobIndexSizeShouldBeTwo_Test(byte heapOffsetSizes, uint result)
         {
             var heapOffsetBasedIndexSizes = new PeNet.Utilities.HeapOffsetBasedIndexSizes(heapOffsetSizes);
 
@@ -65,18 +65,18 @@ namespace PeNet.UnitTest.Utilities
         public void AllIndexesShouldBeFour_Test()
         {
             var heapOffsetBasedIndexSizes = new PeNet.Utilities.HeapOffsetBasedIndexSizes(0x07);
-            Assert.Equal(4, heapOffsetBasedIndexSizes.StringIndexSize);
-            Assert.Equal(4, heapOffsetBasedIndexSizes.GuidIndexSize);
-            Assert.Equal(4, heapOffsetBasedIndexSizes.BlobSize);
+            Assert.Equal((uint) 4, heapOffsetBasedIndexSizes.StringIndexSize);
+            Assert.Equal((uint) 4, heapOffsetBasedIndexSizes.GuidIndexSize);
+            Assert.Equal((uint) 4, heapOffsetBasedIndexSizes.BlobSize);
         }
 
         [Fact]
         public void AllIndexesShouldBeTwo_Test()
         {
             var heapOffsetBasedIndexSizes = new PeNet.Utilities.HeapOffsetBasedIndexSizes(0x00);
-            Assert.Equal(2, heapOffsetBasedIndexSizes.StringIndexSize);
-            Assert.Equal(2, heapOffsetBasedIndexSizes.GuidIndexSize);
-            Assert.Equal(2, heapOffsetBasedIndexSizes.BlobSize);
+            Assert.Equal((uint) 2, heapOffsetBasedIndexSizes.StringIndexSize);
+            Assert.Equal((uint) 2, heapOffsetBasedIndexSizes.GuidIndexSize);
+            Assert.Equal((uint) 2, heapOffsetBasedIndexSizes.BlobSize);
         }
     }
 }
